+++ tools/python/boutiques/zenodoHelper.py
@@ -218,3 +218,35 @@ class ZenodoHelper(object):
             print_info("{0} published to Zenodo, doi is {1}".
                        format(msg_obj, r.json()['doi']), r)
         return r.json()['doi']
+
+    def zenodo_search(self, query, query_line):
+        # Get all results
+        r = requests.get(self.zenodo_endpoint + '/api/records/?q='
+                         'keywords:(/Boutiques/) AND '
+                         'keywords:(/schema-version.*/)'
+                         '%s'
+                         '&file_type=json&type=software&'
+                         'page=1&size=%s' % (query_line, 9999))
+        if(r.status_code != 200):
+            raise_error(ZenodoError, "Error searching Zenodo", r)
+        if(self.verbose):
+            print_info("Search successful for query \"%s\"" % query, r)
+        return r
+
+    def zenodo_upload_file(self, deposition_id, file_path,
+                           zenodo_access_token=None,
+                           error_msg="Cannot Upload to Zenodo",
+                           verbose_msg="Uploaded to Zenodo"):
+        zenodo_access_token = self.get_zenodo_access_token if\
+            zenodo_access_token is None else zenodo_access_token
+        r = requests.post(self.zenodo_endpoint +
+                          '/api/deposit/depositions/%s/files'
+                          % deposition_id,
+                          params={'access_token': zenodo_access_token},
+                          data={'filename': os.path.basename(file_path)},
+                          files={'file': open(file_path, 'rb')})
+
+        if(r.status_code != 201):
+            raise_error(ZenodoError, error_msg, r)
+        if(self.verbose):
+            print_info(verbose_msg, r)
